+++ HISTORY.md
@@ -2,6 +2,13 @@ RELEASE HISTORY
 ===============
 ### Be sure to update the version number in 'runme.py'!
 
+1.0 (2020-04-10)
+-----------------
+  * Integrated chemdescriptor module
+  * Generalized feature report around type_command.csv table
+  * Added full dataframe export option for v3 ETL targets
+  * 
+
 0.86 (2020-03-30)
 -----------------
   * separated baseline parsing from all calcs and feats